requirements.txt
# Minimal requirements for the listener
pyzk
python-dotenv
pyodbc
